+++ lib/Unicode/Char.hs
@@ -45,7 +45,7 @@ import Data.Char (chr, ord)
 import Data.Version (Version, makeVersion)
 import Unicode.Char.Case
 import Unicode.Char.Case.Compat
-import Unicode.Char.General
+import Unicode.Char.General hiding (isLetter, isSpace)
 import Unicode.Char.General.Compat
 import Unicode.Char.Identifiers
 import Unicode.Char.Numeric